[chore] gateway 환경 변수 설정 및 적용
