+++ gateway/src/main/java/com/springcloud/client/gateway/config/PropertyConfig.java
@@ -1,0 +1,11 @@
+package com.springcloud.client.gateway.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.PropertySource;
+import org.springframework.context.annotation.PropertySources;
+
+@Configuration
+@PropertySources({@PropertySource("classpath:env.properties")})
+public class PropertyConfig {
+
+}
