+++ gateway/src/main/resources/application.yml
@@ -30,9 +30,9 @@ spring:
 eureka:
   client:
     service-url:
-      defaultZone: http://localhost:19090/eureka/  # Eureka 서버의 URL을 지정
+      defaultZone: http://${Eureka_name}:${Eureka_password}@${Eureka_host}:19090/eureka/
 
 
 service:
   jwt:
-    secret-key: "401b09eab3c013d4ca54922bb802bec8fd5318192b0a75f201d8b3727429080fb337591abd3e44453b954555b7a0812e1081c39b740293f765eae731f5a65ed1"
+    secret-key: ${Jwt_secret_key}
